requirements.txt
sqlalchemy
python-dotenv
bcrypt